+++ src/main/java/ru/job4j/chat/controller/PersonControl.java
@@ -1,23 +1,37 @@
 package ru.job4j.chat.controller;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.chat.enity.Person;
 import ru.job4j.chat.service.PersonService;
 
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
 
 @RestController
 @RequestMapping("/person")
 public class PersonControl {
-    private final PersonService service;
-    private BCryptPasswordEncoder encoder;
+    private static final Logger LOGGER = LoggerFactory
+            .getLogger(PersonControl.class.getSimpleName());
 
-    public PersonControl(PersonService service, BCryptPasswordEncoder encoder) {
+    private final PersonService service;
+    private final BCryptPasswordEncoder encoder;
+    private final ObjectMapper objectMapper;
+
+    public PersonControl(PersonService service,
+                         BCryptPasswordEncoder encoder,
+                         ObjectMapper objectMapper) {
         this.service = service;
         this.encoder = encoder;
+        this.objectMapper = objectMapper;
     }
 
     @GetMapping("/")
@@ -27,15 +41,29 @@ public class PersonControl {
 
     @GetMapping("/{id}")
     public ResponseEntity<Person> messageById(@PathVariable int id) {
+        if (id <= 0) {
+            throw new NullPointerException("Id cannot be negative or equally 0");
+        }
         var message = service.showById(id);
         return new ResponseEntity<Person>(
-                message.orElse(new Person()),
+                message.orElseThrow(() -> new ResponseStatusException(
+                        HttpStatus.NOT_FOUND,
+                        "Person not found"
+                )),
                 message.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
         );
     }
 
     @PostMapping("/sign-up")
     public ResponseEntity<Person> create(@RequestBody Person person) {
+        if (person.getUsername().isEmpty() || person.getPassword().isEmpty()) {
+            throw new NullPointerException("Username and password mustn't be empty");
+        }
+        if (person.getPassword().length() < 6) {
+            throw new IllegalArgumentException(
+                    "Invalid password. Password length must be more than 5 characters."
+            );
+        }
         person.setPassword(encoder.encode(person.getPassword()));
         return new ResponseEntity<Person>(
                 this.service.save(person),
@@ -45,13 +73,36 @@ public class PersonControl {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Person person) {
+        if (person.getUsername().isEmpty() || person.getPassword().isEmpty()) {
+            throw new NullPointerException("Username and password mustn't be empty");
+        }
+        if (person.getPassword().length() < 6) {
+            throw new IllegalArgumentException(
+                    "Invalid password. Password length must be more than 5 characters."
+            );
+        }
         this.service.save(person);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (id <= 0) {
+            throw new NullPointerException("Id cannot be negative or equally 0");
+        }
         this.service.delete(id);
         return ResponseEntity.ok().build();
     }
+
+    @ExceptionHandler(value = {IllegalArgumentException.class})
+    public void exceptionHandler(Exception e, HttpServletResponse response) throws IOException {
+        response.setStatus(HttpStatus.BAD_REQUEST.value());
+        response.setContentType("application/json");
+        response.getWriter().write(objectMapper.writeValueAsString(new HashMap<>() {{
+            put("massage", e.getMessage());
+            put("type", e.getClass());
+        }}));
+        LOGGER.error(e.getLocalizedMessage());
+    }
+
 }
